Task_3/requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
pyjwt
Flask_Migrate
Flask_JWT_Extended
bcrypt